fix: add suspense for client component
